모든 length 속성에 .toInt() 적용
- List<int>, Uint8List, List 타입의 length 속성이 num을 반환하는 문제 해결
- 모든 경우에 .toInt()를 명시적으로 적용하여 타입 안전성 확보
- 자동 단가 계산 기능이 완전히 안정적으로 작동

+++ lib/features/post_system/screens/post_place_screen.dart
@@ -187,11 +187,11 @@ class _PostPlaceScreenState extends State<PostPlaceScreen> {
     // 사운드 크기 계산
     if (_selectedSound != null) {
       if (_selectedSound is List<int>) {
-        totalSize += _selectedSound.length;
+        totalSize += _selectedSound.length.toInt();
       } else if (_selectedSound is Uint8List) {
-        totalSize += _selectedSound.length;
+        totalSize += _selectedSound.length.toInt();
       } else if (_selectedSound is List) {
-        totalSize += _selectedSound.length;
+        totalSize += _selectedSound.length.toInt();
       }
     }
